Move login and signup validations
